Makefile: Rename manifest.yml to manifest.master.yml to fix loading on fylr
fylr with a plugin include configured on the top level of this repo will fail to start, because the is a manifest.yml on top level and in the build_fylr directory. Fix this by renaming the manifest.yml to manifest.master.yml as we already do in other plugins.

see #78211
see #77710